% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carriots_analytics_ml.R
\name{learn.ca}
\alias{learn.ca}
\title{Default LEARN mechanism of CA datasources}
\usage{
learn.ca()
}
\description{
Reads the data and build the models
###############################################################################
}
