% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carriots_analytics_ml.R
\name{score.ca}
\alias{score.ca}
\title{Default SCORE mechanism of CA datasources}
\usage{
score.ca()
}
\description{
Reads learnt models and score the datasource
###############################################################################
}
